SymSign: decrypt hdr to get pn instead of packet hash
I previously used a packet hash to map encrypted packet with the
associated tag given in the MC_AUTH frames.

It is much simpler to decrypt the packet header only to get the packet
number (pn) to map with the received MC_AUTH tag. Perfs should be
improved.

Thanks to Maxime Piraux and François Michel for the great idea