Soft-multicast uses probe path functions
Soft-multicast option.
     If set alongside the `--multicast` flag, uses multicast QUIC with
     unicast delivery. This option enables to use multicast QUIC without
     IP multicast (or equivalent) support. The multicast QUIC packets are
     delivered on the unicast address given by the client, using the
     multicast port advertised in the MC_ANNOUNCE data.

     For the time beeing, use the `is_ipv6` field of the MC_ANNOUNCE data
     frame to advertise the client if soft multicast is used. This is
     necessary so that the client knows which address they should listen to.

This extensions required to change the way the multicast paths are
created. Before this commit, the client "emulated" the creation of the
path, i.e., without sending a PATH_CHALLENGE frame to the (unicast)
server. This could cause a problem with NAT for examples. Now the
client sends the multipath-related frames to the (unicast) server, even
if neither of them use these additional paths for sending packets.