Add stream_recv_ooo for stream rotation
To prevent applications from reading streams out of order without
knowing it, stream_recv is disabled when the receiving side of the
stream is marked with stream rotation. Instead, applications can use
stream_recv_ooo to receive data out-of-order. In addition, this function
also returns the initial offset of the stream.