Flexicast HTTP/3 with stream rotation
Tests for HTTP/3 over Flexicast quiche work.
The second client, joining the channel after the start of the
transmission, receives all data out of order. It then receives an
Event::Finished.

The problem was the grease frames. This will be a problem if some
clients do not use grease because they will have a different offset from
the flexicast source. This has to be tested but I don't know if it will
work. We will see. In the future.