Better (and correct) handle of receiver fallback
I had to introduce new concepts, such as the "get missing ranges up to"
to correctly enable/disable the handling of acknowledgments for
receivers which fell back on unicast and then came back to the flow,
otherwise "correct" receivers facing losses would not be considered, and
we falsely believe that packets are received on all receivers.

There is still an issue when there are no more receivers in the flow,
because we cannot advance the flexicast flow.