Add mechanism when all receivers fall back
This is a quite strange solution. But when all receivers attached to a leaf controller fall back on unicast, the leaf controller sends dummy acknowledgments, acknowledging all even packets, to be sure that the flexicast flow can continue sending data. This is not safe in case the receivers who fell back cannot follow the pace, but I think it should be ok since the congestion window will be quite small